An AI puzzle-game environment "Temple of Ech"
A frogger-style game designed for AI agents to iterate through with text-map outputs facilitating a "UserInput" agent to participate allowing the observer to play along and try to beat the AI.  Environment files are a collaboration with Ryan Jones, and AgentPatientRun is his independent work.  AgentBestOfBoth combines both his agent and AgentSafeZoneScout.